LayoutTests/media/video-plays-past-end-of-test.html

<p>This test intentionally lets the video keep playing past endTest() to ensure
that shutdown is clean, since DumpRenderTree used to crash in this case.</p>

<video autoplay id="video"></video>
<script src=../http/tests/media/media-file.js></script>
<script src=../http/tests/media/video-test.js></script>
<script type="text/javascript">
if (window.layoutTestController)
  layoutTestController.dumpAsText();
waitForEvent("playing", null, true);
video.src = findMediaFile("video", "content/test");
</script>
