safari-534.53-branch/LayoutTests/media/video-plays-past-end-of-test-expected.txt

This test intentionally lets the video keep playing past endTest() to ensure that shutdown is clean, since DumpRenderTree used to crash in this case.

EVENT(playing)
END OF TEST
